requirements.txt
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.1
matplotlib==3.8.0
scikit-learn==1.3.1
Pillow==10.0.1
tqdm==4.66.1
pyyaml==6.0.1
tensorboard==2.14.0
